+++ kmemory/src/commonMain/kotlin/com/cws/kmemory/FastBuffer.kt
@@ -2,12 +2,12 @@ package com.cws.kmemory
 
 interface FastBuffer {
 
-    var position: Int
-    val capacity: Int
+    var position: Long
+    val capacity: Long
 
     fun release()
     fun getBuffer(): Any
-    fun resize(newCapacity: Int)
+    fun resize(newCapacity: Long)
     operator fun set(index: Int, value: Byte)
     operator fun get(index: Int): Byte
     fun setBytes(index: Int, bytes: ByteArray)
